[PHP 8.1] Move ConstantListClassToEnumRector to PHP 8.1
