+++ phpstan.neon
@@ -705,13 +705,13 @@ parameters:
             message: '#Class has a static method must so must contains "Static" in its name#'
             path: rules/CodingStyle/Enum/PreferenceSelfThis.php
 
-        # optional rule for PHP 8.0
-        - '#Register "Rector\\Php80\\Rector\\Class_\\ConstantListClassToEnumRector" service to "php80\.php" config set#'
-        - '#Rule Rector\\Php80\\Rector\\Class_\\ConstantListClassToEnumRector must implements Rector\\VersionBonding\\Contract\\MinPhpVersionInterface#'
+        # optional rule for PHP 8.1
+        - '#Register "Rector\\Php81\\Rector\\Class_\\ConstantListClassToEnumRector" service to "php81\.php" config set#'
+        - '#Rule Rector\\Php81\\Rector\\Class_\\ConstantListClassToEnumRector must implements Rector\\VersionBonding\\Contract\\MinPhpVersionInterface#'
         - '#Register "Rector\\DowngradePhp80\\Rector\\Enum_\\DowngradeEnumToConstantListClassRector" service to "downgrade\-php80\.php" config set#'
         -
             message: '#Method "refactor(Params|Return)\(\)" returns bool type, so the name should start with is/has/was#'
-            path: rules/Php80/Rector/Class_/ConstantListClassToEnumRector.php
+            path: rules/Php81/Rector/Class_/ConstantListClassToEnumRector.php
 
         # known values
         -
