--- rules-tests/Php80/Rector/Class_/ConstantListClassToEnumRector/Fixture/change_return_type.php.inc
@@ -1,34 +1,0 @@
-<?php
-
-namespace Rector\Tests\Php80\Rector\Class_\ConstantListClassToEnumRector\Fixture;
-
-use Rector\Tests\Php80\Rector\Class_\ConstantListClassToEnumRector\Source\Gear;
-
-final class ChangeReturnType
-{
-    /**
-     * @return Gear::* $gear
-     */
-    public function changeGear(): string
-    {
-        return Gear::FIRST;
-    }
-}
-
-?>
------
-<?php
-
-namespace Rector\Tests\Php80\Rector\Class_\ConstantListClassToEnumRector\Fixture;
-
-use Rector\Tests\Php80\Rector\Class_\ConstantListClassToEnumRector\Source\Gear;
-
-final class ChangeReturnType
-{
-    public function changeGear(): \Rector\Tests\Php80\Rector\Class_\ConstantListClassToEnumRector\Source\Gear
-    {
-        return Gear::FIRST;
-    }
-}
-
-?>
