--- rules-tests/Php80/Rector/Class_/ConstantListClassToEnumRector/Fixture/enum_type_property.php.inc
@@ -1,28 +1,0 @@
-<?php
-
-namespace Rector\Tests\Php80\Rector\Class_\ConstantListClassToEnumRector\Fixture;
-
-use Rector\Tests\Php80\Rector\Class_\ConstantListClassToEnumRector\Source\Gear;
-
-final class EnumTypeProperty
-{
-    /**
-     * @var Gear::* $gear
-     */
-    private $gear;
-}
-
-?>
------
-<?php
-
-namespace Rector\Tests\Php80\Rector\Class_\ConstantListClassToEnumRector\Fixture;
-
-use Rector\Tests\Php80\Rector\Class_\ConstantListClassToEnumRector\Source\Gear;
-
-final class EnumTypeProperty
-{
-    private \Rector\Tests\Php80\Rector\Class_\ConstantListClassToEnumRector\Source\Gear $gear;
-}
-
-?>
