+++ rules-tests/Php81/Rector/Class_/ConstantListClassToEnumRector/Fixture/change_param_type.php.inc
@@ -1,0 +1,32 @@
+<?php
+
+namespace Rector\Tests\Php81\Rector\Class_\ConstantListClassToEnumRector\Fixture;
+
+use Rector\Tests\Php81\Rector\Class_\ConstantListClassToEnumRector\Source\Gear;
+
+final class ChangeParamType
+{
+    /**
+     * @param Gear::* $gear
+     */
+    public function changeGear(string $gear)
+    {
+    }
+}
+
+?>
+-----
+<?php
+
+namespace Rector\Tests\Php81\Rector\Class_\ConstantListClassToEnumRector\Fixture;
+
+use Rector\Tests\Php81\Rector\Class_\ConstantListClassToEnumRector\Source\Gear;
+
+final class ChangeParamType
+{
+    public function changeGear(\Rector\Tests\Php81\Rector\Class_\ConstantListClassToEnumRector\Source\Gear $gear)
+    {
+    }
+}
+
+?>
