+++ rules-tests/Php81/Rector/Class_/ConstantListClassToEnumRector/Fixture/skip_non_public_const.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\Tests\Php80\Rector\Class_\ConstantListClassToEnumRector\Fixture;
+namespace Rector\Tests\Php81\Rector\Class_\ConstantListClassToEnumRector\Fixture;
 
 final class SkipNonPublicConst
 {
