+++ rules-tests/Php81/Rector/Class_/ConstantListClassToEnumRector/Fixture/some_class.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\Tests\Php80\Rector\Class_\ConstantListClassToEnumRector\Fixture;
+namespace Rector\Tests\Php81\Rector\Class_\ConstantListClassToEnumRector\Fixture;
 
 class Direction
 {
@@ -13,7 +13,7 @@ class Direction
 -----
 <?php
 
-namespace Rector\Tests\Php80\Rector\Class_\ConstantListClassToEnumRector\Fixture;
+namespace Rector\Tests\Php81\Rector\Class_\ConstantListClassToEnumRector\Fixture;
 
 enum Direction : string
 {
