+++ rules-tests/Php81/Rector/Class_/ConstantListClassToEnumRector/Source/Gear.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Rector\Tests\Php80\Rector\Class_\ConstantListClassToEnumRector\Source;
+namespace Rector\Tests\Php81\Rector\Class_\ConstantListClassToEnumRector\Source;
 
 final class Gear
 {
