+++ rules-tests/Php81/Rector/Class_/ConstantListClassToEnumRector/config/configured_rule.php
@@ -3,7 +3,7 @@
 declare(strict_types=1);
 
 use Rector\Config\RectorConfig;
-use Rector\Php80\Rector\Class_\ConstantListClassToEnumRector;
+use Rector\Php81\Rector\Class_\ConstantListClassToEnumRector;
 
 return static function (RectorConfig $rectorConfig): void {
     $rectorConfig->rule(ConstantListClassToEnumRector::class);
