+++ rules/Php81/NodeAnalyzer/EnumConstListClassDetector.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Rector\Php80\NodeAnalyzer;
+namespace Rector\Php81\NodeAnalyzer;
 
 use PhpParser\Node\Scalar;
 use PhpParser\Node\Stmt\Class_;
