+++ rules/Php81/Rector/Class_/ConstantListClassToEnumRector.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Rector\Php80\Rector\Class_;
+namespace Rector\Php81\Rector\Class_;
 
 use PhpParser\Node;
 use PhpParser\Node\Name\FullyQualified;
@@ -16,15 +16,15 @@ use Rector\BetterPhpDocParser\PhpDocInfo\PhpDocInfo;
 use Rector\BetterPhpDocParser\PhpDocManipulator\PhpDocTagRemover;
 use Rector\Core\Rector\AbstractRector;
 use Rector\Core\Reflection\ReflectionResolver;
-use Rector\Php80\NodeAnalyzer\EnumConstListClassDetector;
 use Rector\Php80\NodeAnalyzer\EnumParamAnalyzer;
 use Rector\Php80\ValueObject\ClassNameAndTagValueNode;
+use Rector\Php81\NodeAnalyzer\EnumConstListClassDetector;
 use Rector\Php81\NodeFactory\EnumFactory;
 use Symplify\RuleDocGenerator\ValueObject\CodeSample\CodeSample;
 use Symplify\RuleDocGenerator\ValueObject\RuleDefinition;
 
 /**
- * @see \Rector\Tests\Php80\Rector\Class_\ConstantListClassToEnumRector\ConstantListClassToEnumRectorTest
+ * @see \Rector\Tests\Php81\Rector\Class_\ConstantListClassToEnumRector\ConstantListClassToEnumRectorTest
  */
 final class ConstantListClassToEnumRector extends AbstractRector
 {
